requirements.txt
pandas
numpy
matplotlib
scikit-learn
spacy
PyPDF2
python-docx
fpdf